Servant.Swagger.Internal.TypeLevel.API

Documentation

type family EndpointsList api

Build a list of endpoints from an API.

Equations

EndpointsList (a :<|> b) = AppendList (EndpointsList a) (EndpointsList b)

EndpointsList (e :> a) = MapSub e (EndpointsList a)

EndpointsList a = `[a]`

type family IsSubAPI sub api :: Constraint

Check whether sub is a sub API of api.

Equations

IsSubAPI sub api = AllIsElem (EndpointsList sub) api

type family AllIsElem xs api :: Constraint

Check that every element of xs is an endpoint of api.

Equations

AllIsElem `[]` api = ()

AllIsElem (x : xs) api = (IsIn x api, AllIsElem xs api)

type family MapSub e xs

Apply (e :>) to every API in xs.

Equations

MapSub e `[]` = `[]`

MapSub e (x : xs) = (e :> x) : MapSub e xs

type family AppendList xs ys

Append two type-level lists.

Equations

AppendList `[]` ys = ys

AppendList (x : xs) ys = x : AppendList xs ys

type family Or a b :: Constraint

Equations

Or () b = ()

Or a () = ()

type family IsIn sub api :: Constraint

Equations

IsIn e (a :<|> b) = Or (IsIn e a) (IsIn e b)

IsIn (e :> a) (e :> b) = IsIn a b

IsIn e e = ()

type family Elem x xs :: Constraint

Check whether a type is a member of a list of types. This is a type-level analogue of elem.

Equations

Elem x (x : xs) = ()

Elem x (y : xs) = Elem x xs

type family IfThenElse c a b

A type-level analogue of if then else statement.

Equations

IfThenElse () a b = a

IfThenElse c a b = b

type family BodyTypes c api :: [*]

Extract a list of "body" types for a specific content-type from a servant API.

Equations

BodyTypes c (Delete cs (Headers hdrs a)) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Get cs (Headers hdrs a)) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Patch cs (Headers hdrs a)) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Post cs (Headers hdrs a)) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Put cs (Headers hdrs a)) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Delete cs a) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Get cs a) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Patch cs a) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Post cs a) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (Put cs a) = IfThenElse (Elem c cs) `[a]` `[]`

BodyTypes c (ReqBody cs a :> api) = IfThenElse (Elem c cs) (a : BodyTypes c api) (BodyTypes c api)

BodyTypes c (e :> api) = BodyTypes c api

BodyTypes c (a :<|> b) = AppendList (BodyTypes c a) (BodyTypes c b)