Servant.Swagger.Internal.TypeLevel.Every

Documentation

>>> :set -XDataKinds
>>> :set -XFlexibleContexts
>>> :set -XGADTs
>>> :set -XRankNTypes
>>> :set -XScopedTypeVariables
>>> import GHC.TypeLits
>>> import Data.List

type family EveryTF cs x :: Constraint

Apply multiple constraint constructors to a type.

EveryTF '[Show, Read] a ~ (Show a, Read a)

Note that since this is a type family, you have to alway fully apply EveryTF.

For partial application of multiple constraint constructors see Every.

Equations

EveryTF `[]` x = ()

EveryTF (c : cs) x = (c x, EveryTF cs x)

data EveryProof cs x where

EveryProof cs x is a proof that x satisfies all constraints in cs.

E.g. a value of type EveryProof '[Show, Read] a proves that a has both Show and Read instances.

Pattern matching on EveryProofCons brings one constraint at a time to the scope.

Constructors

EveryProofNil :: EveryProof `[]` x

EveryProofCons :: c x => EveryProof cs x -> EveryProof (c : cs) x

class Every cs x where

Apply multiple constraint constructors to a type as a class.

This is different from EveryTF in that it allows partial application. However, Every cs does not immediately bring EveryTF into scope, so you can't use the instances right away:

>>> :{
let f :: Every [Show, Num] a => a -> String
    f = show
:}
...
    Could not deduce (Show a) arising from a use of ‘show’
    from the context (Every '[Show, Num] a)
...

To bring individual constraints into scope use every and pattern match on the proof:

>>> :{
let f :: forall a. Every [Show, Num] a => a -> String
    f = case (every :: EveryProof [Show, Num] a) of EveryProofCons (EveryProofCons EveryProofNil) -> show
in f 123
:}
"123"

Alternatively, you can use appWith to convert Every cs constraint into EveryTF cs. appWith pqs pcs f px applies a function f with regular constraints to a proxied type px, using Every cs x constraint of px. It automatically proves that px satisfies the regular constraints of f.

>>> let num = Proxy :: Proxy '[Num]
>>> let enumOrd = Proxy :: Proxy '[Enum, Ord]
>>> let xs :: (Num a, Enum a, Ord a) => p a -> [a]; xs _ = sort [1, 4, 3, 2]
>>> let ys :: (Num a, Every '[Enum, Ord] a) => p a -> [a]; ys = appWith num enumOrd xs
>>> ys (Proxy :: Proxy Int)
[1,2,3,4]

It might be useful to look at the type of appWith also like this:

appWith :: pq qs -> p cs
   -> (forall x p'. (EveryTF qs x, EveryTF cs x) => p' x -> f x)
   -> (forall x p'. (EveryTF qs x, Every   cs x) => p' x -> f x)

Methods

every :: EveryProof cs x

appWith :: EveryTF qs x => pq qs -> p cs -> (forall y p'. (EveryTF qs y, EveryTF cs y) => p' y -> f y) -> p'' x -> f x

appWith_ :: (EveryTF qs x, Every cs x) => pq qs -> p cs -> (forall y p'. (EveryTF qs y, EveryTF cs y) => p' y -> a) -> p'' x -> a

Convert some constraints from Every to EveryTF for a function which result type does not depend on its argument's type.

app :: Every cs x => p cs -> (forall y p'. EveryTF cs y => p' y -> f y) -> p'' x -> f x

Convert all constraints from Every to EveryTF.

app_ :: Every cs x => p cs -> (forall y p'. EveryTF cs y => p' y -> a) -> p'' x -> a

Convert all constraints from Every to EveryTF for a function which result type does not depend on its argument's type.

appVal_ :: forall p a cs x. Every cs x => p cs -> (forall y. EveryTF cs y => y -> a) -> x -> a

Convert all constraints from Every to EveryTF for a term-level function which result type does not depend on its argument's type.

tmapEvery :: forall a cs p p'' xs. TMap (Every cs) xs => p cs -> (forall x p'. EveryTF cs x => p' x -> a) -> p'' xs -> [a]

Like tmap, but uses Every for multiple constraints.

>>> let zero :: forall p a. (Show a, Num a) => p a -> String; zero _ = show (0 :: a)
>>> tmapEvery (Proxy :: Proxy [Show, Num]) zero (Proxy :: Proxy [Int, Float])
["0","0.0"]